+++ tools/batools/apprun.py
@@ -1,6 +1,6 @@
 # Released under the MIT License. See LICENSE for details.
 #
-"""Utils for wrangling runs of the app.
+"""Utils for wrangling running the app as part of a build.
 
 Manages constructing or downloading it as well as running it.
 """
@@ -105,6 +105,9 @@ def acquire_binary(assets: bool, purpose: str) -> str:
     if os.environ.get('BA_APP_RUN_ENABLE_BUILDS') == '1':
         # Going the build-it-ourselves route.
 
+        # Don't need any env mods for this path.
+        env = None
+
         if os.environ.get('BA_APP_RUN_BUILD_HEADLESS') == '1':
             # User has opted for headless builds.
             if assets:
@@ -141,16 +144,23 @@ def acquire_binary(assets: bool, purpose: str) -> str:
                 binary_build_command = ['make', 'cmake-binary']
             binary_path = 'build/cmake/debug/staged/ballisticakit'
     else:
-        # Ok; going with prefab headless stuff.
+        # Ok; going with a downloaded prefab headless build.
 
-        # Let the user know how to use their own binaries instead.
+        # By default, prefab build targets on WSL (Linux running on
+        # Windows) will give us Windows builds which won't work right
+        # here. Ask it for Linux builds instead.
+        env = dict(os.environ, BA_WSL_TARGETS_WINDOWS='0')
+
+        # Let the user know how to use their own built binaries instead
+        # if they prefer.
         note = '\n' + textwrap.fill(
-            'NOTE: You can set env-var BA_APP_RUN_ENABLE_BUILDS=1'
-            f' to use locally-built binaries for {purpose}'
-            ' instead of prefab ones. This will properly reflect any changes'
-            ' you\'ve made to the C/C++ layer.',
+            f'NOTE: You can set env-var BA_APP_RUN_ENABLE_BUILDS=1'
+            f' to use locally-built binaries for {purpose} instead'
+            f' of prefab ones. This will properly reflect any changes'
+            f' you\'ve made to the C/C++ layer.',
             80,
         )
+
         if assets:
             print(
                 f'{Clr.SMAG}Fetching prefab binary & assets for'
@@ -160,6 +170,7 @@ def acquire_binary(assets: bool, purpose: str) -> str:
             binary_path = (
                 subprocess.run(
                     ['tools/pcommand', 'prefab_binary_path', 'server-release'],
+                    env=env,
                     check=True,
                     capture_output=True,
                 )
@@ -176,6 +187,7 @@ def acquire_binary(assets: bool, purpose: str) -> str:
             binary_path = (
                 subprocess.run(
                     ['tools/pcommand', 'prefab_binary_path', 'server-release'],
+                    env=env,
                     check=True,
                     capture_output=True,
                 )
@@ -184,7 +196,7 @@ def acquire_binary(assets: bool, purpose: str) -> str:
             )
             binary_build_command = ['make', binary_path]
 
-    subprocess.run(binary_build_command, check=True)
+    subprocess.run(binary_build_command, env=env, check=True)
     if not os.path.exists(binary_path):
         raise RuntimeError(
             f"Binary not found at expected path '{binary_path}'."