+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### 1.7.33 (build 21778, api 8, 2024-03-13)
+### 1.7.33 (build 21790, api 8, 2024-03-14)
 - Stress test input-devices are now a bit smarter; they won't press any buttons
   while UIs are up (this could cause lots of chaos if it happened).
 - Added a 'Show Demos When Idle' option in advanced settings. If enabled, the
@@ -40,11 +40,15 @@
   EraOSBeta!)
 - Added a UI for customizing Series Length in Teams and Points-to-Win in FFA
   (Thanks EraOSBeta!)
-- Added a 'Practice' button in tournaments, letting you play them free of
-  charge while not submitting scores. (Thanks Temp!)
+- Implemented HEX code support to the advanced color picker (Thanks 3alTemp!)
+- Players leaving the game after getting hurt will now grant kills. (Thanks
+  Temp!)
 - Sphinx based Python documentation generation is now wired up (Thanks
   Loup-Garou911XD!)
 - Renaming & overwriting existing profiles is no longer possible (Thanks Temp!)
+- Cleaned up builds when running under WSL. Things like `make mypy` should now
+  work correctly there, and it should now be possible to build and run either
+  Linux or Windows builds there.
 
 ### 1.7.32 (build 21741, api 8, 2023-12-20)
 - Fixed a screen message that no one will ever see (Thanks vishal332008?...)
@@ -71,6 +75,8 @@
   cleanly however (an `on_app_active_changed()` call in the `AppMode` class).
   This means that it also applies to other platforms when the app reaches the
   'inactive' state; for instance when minimizing the window on the SDL build.
+- Added an `allow_clear_button` arg to bauiv1.textwidget() which can be used to
+  disable the 'X' button that clears editable text widgets.
 
 ### 1.7.31 (build 21727, api 8, 2023-12-17)
 - Added `bascenev1.get_connection_to_host_info_2()` which is an improved